core/standards/web-apps/media/semantics/preload/metadata/preload-metadata-to-none-after-loadstart-networkstate-webm.html

<!doctype html>
<!-- This file was generated from 'preload-metadata-to-none-after-loadstart-networkstate.tpl' using 'tpl2tc.py'. Please do not edit manually! -->
<title>HTMLMediaElement.preload:metadata networkState when preload is set to none after loadstart is fired - webm</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../support/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="metadata" controls></audio>
<video preload="metadata" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.preload:metadata networkState when preload is set to none after loadstart is fired - webm", "HTMLVideoElement.preload:metadata networkState when preload is set to none after loadstart is fired - webm", {timeout:10000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
    var tag_name = vars[1];
    var states_actual = '';
    var node = document.getElementsByTagName(tag_name)[0];
    if (!node.canPlayType(tag_name + '/webm')) {
        t.step(function() {
        assert_unreached("Cannot play '" + tag_name + "/webm'");
        t.done();
        });
    }
    node.addEventListener('loadstart', function(){ node.preload='none'; }, false);
    node.addEventListener('suspend', endTest, false);
    node.addEventListener('error', function(e) {
        t.step(function() {
        assert_unreached("Error event was fired. Error code: " +  e.target.error.code);
        t.done();
        });
    }, false);
    node.src = '/core/standards/web-apps/media/network/range-request-log/range-request.php?rate=100000&fileloc=../../support/preload.webm&nocache=' + Math.random();

    function endTest() {
        t.step(function() {
        assert_equals(node.networkState, HTMLMediaElement.NETWORK_IDLE, 'networkState after suspend')
        t.done();
        });
    }
    });
});
</script>
